pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wireviz-yaml-generator"
version = "0.1.0"
description = "Generate WireViz YAML files and manufacturing documentation from SQLite electrical design databases"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Ole Johan Bondahl", email = "your.email@example.com"}
]
keywords = [
    "wireviz",
    "wiring-diagram",
    "electrical",
    "harness",
    "yaml",
    "bom",
    "manufacturing"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Manufacturing",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "wireviz>=0.4.1",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "openpyxl>=3.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.9",
    "ty>=0.0.1a0",
]

[project.urls]
Homepage = "https://github.com/OleJBondahl/wireviz_yaml_generator"
Documentation = "https://olejbondahl.github.io/wireviz_yaml_generator/"
Repository = "https://github.com/OleJBondahl/wireviz_yaml_generator"
Issues = "https://github.com/OleJBondahl/wireviz_yaml_generator/issues"

[project.scripts]
wireviz-generator = "wireviz_yaml_generator.main:main"

[tool.hatch.build.targets.wheel]
packages = ["wireviz_yaml_generator"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["wireviz_yaml_generator", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.ty.environment]
python-version = "3.12"
